Finishing tree-breadth-first
