+++ python/code_challenges/breadthfirst/README.md
@@ -1,0 +1,34 @@
+
+# Trees
+<!-- Short summary or background information -->
+A Tree is combination of nodes connected with each other ,  node is a component which may contain it’s own values, and references to other nodes
+
+## Challenge
+<!-- Description of the challenge -->
+Create Breadth-first Traversal for binary tree , which takes in a tree as argument , and return a list of all values in the tree
+
+## Approach & Efficiency
+<!-- What approach did you take? Why? What is the Big O space/time for this approach? -->
+
+* time and complexity for (breadth_first) method: O(n)
+
+
+
+
+## whiteboard for Code Challenge: Class 17
+
+![code17](../../codechallenege17.png)
+
+
+
+
+## API
+<!-- Description of each method publicly available in each of your trees -->
+
+
+
+>breadth_first"""
+    A binary tree method which returns a list of items that it contains
+    Arguments: tree
+    Return: list of all values in the tree, in the order they were encountered
+    """
